deepSORT_requirements.txt
absl-py==0.15.0
asgiref==3.7.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h5py==2.10.0
idna==3.6
image==1.5.32
importlib-metadata==7.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.3.0.38
opt-einsum==3.3.0
pillow==10.2.0
protobuf==3.20.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
six==1.15.0
sqlparse==0.4.4
tensorboard==2.2.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.1
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.12.1
zipp==3.17.0
